demo/requirement.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dateutils==0.6.8
Django==3.0.7
django-dropbox-storage==0.4.1
django-filter==2.3.0
django-restframework==0.0.1
django-storages==1.9.1
djangorestframework==3.11.0
drf-yasg==1.17.1
dropbox==10.2.0
idna==2.10
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.4
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
